+++ src/notices/types.d.ts
@@ -1,0 +1,7 @@
+export type PaymentWarningModeType = "home" | "overview" | "billing";
+
+export interface PaymentMethodError {
+  type: number,
+  orgaName?: string,
+  orgaBillingLink?: string,
+}