+++ test/i18n-sanitize.test.js
@@ -42,43 +42,43 @@ describe('filter', () => {
     );
   });
 
-  it('Whitelist <strong>', () => {
+  it('Authorize <strong>', () => {
     compareChildNodes(sanitize`One <strong>Two</strong> Three Four Five Six`,
       ['One ', ['strong', ['Two']], ' Three Four Five Six'],
     );
   });
 
-  it('Whitelist <em>', () => {
+  it('Authorize <em>', () => {
     compareChildNodes(sanitize`One Two <em>Three</em> Four Five Six`,
       ['One Two ', ['em', ['Three']], ' Four Five Six'],
     );
   });
 
-  it('Whitelist <code>', () => {
+  it('Authorize <code>', () => {
     compareChildNodes(sanitize`One Two Three <code>Four</code> Five Six`,
       ['One Two Three ', ['code', ['Four']], ' Five Six'],
     );
   });
 
-  it('Whitelist <a>', () => {
+  it('Authorize <a>', () => {
     compareChildNodes(sanitize`One Two Three Four <a>Five</a> Six`,
       ['One Two Three Four ', ['a', ['Five']], ' Six'],
     );
   });
 
-  it('Replace non-whitelisted tags with a text node and escape the contents', () => {
+  it('Replace non-authorized tags with a text node and escape the contents', () => {
     compareChildNodes(sanitize`One <div><strong>Two</strong></div> <pre><em>Three</em></pre> <h1><code>Four</code></h1> <div><code>Five</code></div> <section><a>Six</a></section>`,
       ['One Two Three Four Five Six'],
     );
   });
 
-  it('Remove all attributes on whitelisted tags', () => {
+  it('Remove all attributes on authorized tags', () => {
     compareChildNodes(sanitize`One <strong href="/foo" onclick="alert('xss')">Two</strong> <em foo-attribute="bar">Three</em> <code class="the-class">Four</code> <a>Five</a> Six`,
       ['One ', ['strong', ['Two'], []], ' ', ['em', ['Three'], []], ' ', ['code', ['Four'], []], ' ', ['a', ['Five'], []], ' Six'],
     );
   });
 
-  it('Keep title attribute on whitelisted tags', () => {
+  it('Keep title attribute on authorized tags', () => {
     compareChildNodes(sanitize`One <strong title="strong-title">Two</strong> <em title="em-title">Three</em> <code title="code-title">Four</code> <a title="a-title">Five</a> Six`,
       [
         'One ',
@@ -96,7 +96,7 @@ describe('filter', () => {
 
   it('Only keep href and title attributes on <a> (remove every other attributes)', () => {
     compareChildNodes(sanitize`One Two Three Four <a href="/foobar" download id="the-id" target="foobar" title="a-title">Five</a> Six`,
-      ['One Two Three Four ', ['a', ['Five'], [['href', '/foobar'], ['title', 'a-title']]], ' Six'],
+      ['One Two Three Four ', ['a', ['Five'], [['href', '/foobar'], ['title', 'a-title'], ['class', 'sanitized-link']]], ' Six'],
     );
   });
 
@@ -106,7 +106,7 @@ describe('filter', () => {
     );
   });
 
-  it('Everything that comes from params must be escaped (even whitelisted tags)', () => {
+  it('Everything that comes from params must be escaped (even authorized tags)', () => {
     compareChildNodes(sanitize`${'<h1>One</h1>'} ${'<strong>Two</strong>'} ${'<em>Three</em>'} ${'<code>Four</code>'} ${'<a>Five</a>'} ${'<pre>Six</pre>'}`,
       ['<h1>One</h1> <strong>Two</strong> <em>Three</em> <code>Four</code> <a>Five</a> <pre>Six</pre>'],
     );